moving files and congregating
