+++ src/components/plots/GeneReorderTool.jsx
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 
 import { updatePlotConfig } from 'redux/actions/componentConfig';
 
-import { arrayMoveImmutable } from 'utils/array-move';
+import { arrayMoveImmutable } from 'utils/arrayUtils';
 import HierarchicalTreeGenes from 'components/plots/hierarchical-tree-genes/HierarchicalTreeGenes';
 
 import { Space, Button } from 'antd';
@@ -14,7 +14,7 @@ const GeneReorderTool = (props) => {
   const { plotUuid, onDelete } = props;
 
   const dispatch = useDispatch();
-  
+
   const config = useSelector((state) => state.componentConfig[plotUuid]?.config);
 
   // Tree from antd requires format [{key: , title: }], made from gene names from loadedMarkerGenes and config
