+++ src/pages/data-management/index.jsx
@@ -5,8 +5,8 @@ import { loadExperiments } from 'redux/actions/experiments';
 import Header from 'components/Header';
 import MultiTileContainer from 'components/MultiTileContainer';
 import NewProjectModal from 'components/data-management/NewProjectModal';
-import ProjectsListContainer from 'components/data-management/ProjectsListContainer';
-import ProjectDetails from 'components/data-management/ProjectDetails';
+import ProjectsListContainer from 'components/data-management/project/ProjectsListContainer';
+import ProjectDetails from 'components/data-management/project/ProjectDetails';
 import { loadProcessingSettings } from 'redux/actions/experimentSettings';
 import loadBackendStatus from 'redux/actions/backendStatus/loadBackendStatus';
 import { loadSamples } from 'redux/actions/samples';
