+++ src/redux/actions/componentConfig/downloadNormalizedMatrix.js
@@ -1,9 +1,9 @@
 import { PLOT_DATA_LOADED, PLOT_DATA_LOADING, PLOT_DATA_ERROR } from 'redux/actionTypes/componentConfig';
 
-import downloadFromUrl from 'utils/downloadFromUrl';
+import downloadFromUrl from 'utils/upload/downloadFromUrl';
 import handleError from 'utils/http/handleError';
 import fetchWork from 'utils/work/fetchWork';
-import writeToFileURL from 'utils/writeToFileURL';
+import writeToFileURL from 'utils/upload/writeToFileURL';
 
 const downloadNormalizedMatrix = (
   plotUuid,
