+++ src/utils/extraActionCreators/downloadProcessedMatrix.js
@@ -3,7 +3,7 @@ import { loadProcessingSettings } from 'redux/actions/experimentSettings';
 
 import fetchWork from 'utils/work/fetchWork';
 import getTimeoutForWorkerTask from 'utils/getTimeoutForWorkerTask';
-import writeToFileURL from 'utils/writeToFileURL';
+import writeToFileURL from 'utils/upload/writeToFileURL';
 import downloadFromUrl from 'utils/downloadFromUrl';
 import getEmbeddingETag from 'utils/work/getEmbeddingETag';
 import handleError from 'utils/http/handleError';
